If release notes will be available later, tweet when they are
Apple may say that release notes for an update will be 'available soon'. Now the bot will save this releases and next time check if release notes are now available. If yes, it will tweet that in a new tweet.
This data and zero day CVEs are now stored in a json file