feat: store tweeted releases in stored_data.json
For example, if the bot tweets new releases and Apple makes another release a few minutes after that, it causes issues as when executed again the bot will tweet all of the previous releases again from that day. This is now changed and the bot will save current date and all of the tweeted releases from the current day in stored_data.json file - next time they get overwritten. If the bot is ran twice on the same day it will only tweet releases that were not tweeted yet.