fix: not catching changes between 11pm and 12pm
As the bot is running hourly and always checking with the current date format, it did not catch changes that were made between 11pm and 12pm. Now if todays_date is incorrect, read_file() returns that it is midnight. With midnight being True bot is doing checks with the previous date. Updating todays_date and reseting stored_data for the new day is now in the save_file function after the midnight check.