fix: receiving no_details_yet release twice tweeted security_content_available
At the first time, bot would save the release name to the stored_data.json and delete it from being tweeted. The next hour, no if statement was there to catch the same release coming in twice and it got tweeted out as if the security content was available. This commit also adds a test that can catch this in the future.