feat: re-format how Emergency Updates zero-day post looks like
Previously if there were many zero-days it was unclear which were new
and which were old.

Previously if there were many releases, it would say zero-day details
and in the next thread post it would say "1 zero-day fixed in x" and
"3 zero-days fixed in y". This looks good but does not tell you which
zero-days are in x and which in y, just says the number.

Previously if there was only one zero-day patch update it would make
two thread posts.

This issues are being solved in this commit. It makes data more easily
readable and more clear what is being fixed in each release.