feat: experimental bot hourly runs inside of GitHub Actions
I have set up a GitHub cron job that will run the bot hourly as an
experiment.

As I store data in a JSON file, this file has to be shared between runs.
I *think* I have achieved this with Artifacts.

If this works as expected, this will become the sole place for this bot.

Currently it will only post to my test accounts as I do not
trust it yet :D