Initial commit for the posted_data_json branch
This branch will be used for storing posted_data.json file.

This will enable the GitHub Action bot to quickly read and write data
to it. Will allow for seeing all of the changes made by the bot.
Alongside it will also allow me to easily modify it and be sure the bot
will get my changes next time.