fix: skip release if its security content page parsing fails
Currently there are 3 old releases that are causing trouble:
- tvOS 14.6
- iCloud for Windows 11.3
- iTunes 12.10.8 for Windows

For some reason, the bot gets redirected to their Arabic sites at
every midnight run (when its doing a daily entry changes check).

I wanted to skip these based on titles, but ended with try/catch
as this is unpredictable and the failure should not be breaking.