feat: catch releases for which Apple sets a non-today release date
On March 7, 2023 Apple released "GarageBand for macOS 10.4.8". Security
content of this releases did not become available until March 9. On this
date Apple added the release and set its release date on March 7.

The bot was only checking releases that were released on the current
day. So on March 9, there were no posts made.

This commit should resolve this. Bot now instead checks the name of the
release and sees if it is in the posted_data.json

This file previously stored data posted on the current day and was reset
on midnight. Now it stores x number of already posted release names.

The bot checks if release is in there, else posts it. No longer relies
on the date.