ref: better type annotations, access variables without getters
Improved type annotations for list and dict types with more specifics
of what they contain.

Previously there were custom getters, but with no custom behavior as
there are custom setters and no need for anyone else to change these
variables.

This commit converts those getters to a @property. This allows for
normal reading of variables as release.name but does not allow setting
values to them.