ref: Release class
Change setters to retrieve_x() as that seems more what they do. These
functions are only called at the beginning and they set the Release
attributes.

Some of the setters modified multiple attributes, so logic was not
very separated. These functions are now also static, so they cannot
even change class attributes but only return values.

So the only place the Release attributes now get changed is inside of the
__init__()