fix: also catch changes in "Additional recognition" section
When HTML parsing got moved to more lxml in November 2022, it has been
set to strip "Additional recognition" title and everything after it.

As only the Entry Changes function needs that section, this went
unnoticed for a while.

So from now on, Entry Changes should also check and count changes
inside of Additional recognition.