fix: add fail-safes for posted_data.json[posts][new_releases] list
Recently, the bot moved from checking release dates to checking release
names that have not yet been posted. This move was motivated as Apple
added a new release with release date set to a few days ago, causing
the bot to not catch it.

To know which releases have been posted, bot keeps a list inside of
posted_data.json file. It goes through Apple's website and when it
stumbles upon the first release that is in this list, it stops.

Now, if the list is empty, it will think all of the releases are new.
A new check has been added for this.

If list is not empty, but for some reason none of the release names
are found, it will now fail if it finds more than 20 "new" releases.