ref+tests: separate parsing in Release class, remove ReleaseTest
Previously most of parsing was done in __init__, which was probably
not of one the best ideas. All of the parsing has now been moved to
static methods inside of Release.

Now there is a clean __init__, which can be used in tests. This
completely removes the need for ReleaseTest.