ref: distinction between "new security content" and "new release"
This is a preparation for the next changes.

Previously in the code "new release" or "new update" was considered
when Apple Security page was updated with the security content.

This is no longer the case. This is now "new security content" or similar.

Post will now also say "NEW SECURITY DETAILS" instead of "NEW UPDATES".

This commit also merges a lot of posting code of (previous)
"NEW UPDATES" and "SECURITY CONTENT AVAILABLE" (when Apple said
"details soon") into one.