ref: for now do not post to main Mastodon account
Particularly I am not sure about the checking if release was already
posted or not, as Apple removes old releases from RSS when new ones
come. I might need to move to checking the dates instead?

We will see today :)

And about Twitter, the bot has not been posting directly to the main
account for a while now.