feat: in zero-days post group CVEs if they fixed in same set of releases
Currently each CVE zero-day is listed in a post separately and
below it has listed releases in which it was fixed.

The issue is if there are 3 zero-days all fixed in same 4 releases,
each CVE would be separate and repeat same 4 releases.

With this commit multiple CVEs can be together, separated by
comma and then have only one listing of releases, if it is the same.