fix: for new RSS releases rely on date and not on already_posted names
Initially this was only checking if release was already posted. All
releases in RSS read before one of the names appeared in PostedDataFile,
were considered new.

The problem is, that Apple removes old releases from RSS, so releases
in PostedDataFile might no longer be in the RSS feed, causing bot
to consider wrong releases new.

For example when "iOS 17 beta 2" is released, its predecessor is
removed from RSS.

So this commit removes this reliance on already posted names and
moves to checking the published date first and then if it was already
posted.