feat: add security content links below release names in some posts
Previously Entry Changes contained no links and New Security Content
had a link at the end to the main security page, prompting the user
to search manually if he wanted to see security content of a said release.

This is no longer the case for those two posts.

Disabled for Twitter for now as it makes posts too long and then just
makes a really long thread of tweets.

But on Mastodon, where there is a 11k character limit, this is great!