feat: sort releases alphabetically in zero-day post for each CVE
On 21 Jun 2023, Apple pushed updates for 3 new zero-days. One of the
zero-days was patched in 7 releases with updates for iOS 16, iOS 15,
multiple watchOS and macOS versions.

Systems were scattered around, one there and one there, making
it hard to quickly grasp the list and see the latest versions of each one.