feat: post directly to the main account
A few updates have now gone through and everything is at it should be,
so this is no longer needed.

Posting to test account is only enabled for posts that post new
releases, read from Apple RSS feed. I am not sure if this is how I want
it to work yet, or only limit it to beta updates.

(Twitter API is still working - at least the test account one)