fix: no longer stop if posting to Twitter fails, it is unreliable
Previously if there were issues with publishing the post, the bot would
fail, I would see it and fix it.

But with Twitter issues are regular and it is no longe reliable enough
to have the bot stop on post fail.

The bot will now only stop if posting to Mastodon fails, where as of now
there were zero posting issues.