feat: improve counting of zero-days and module bugs
In the "fixed in" post the list of modules was not really accurate
as it was counting the number of title occurrences and not CVEs inside
under them.

Move this fix to zero-days too. Apple says that something is a zero-day
in the Impact section and *can* then list multiple CVEs for it below.
It always picked first and I am not aware of any cases where there were
multiple CVEs together in one entry but still.