ref: optimize "sec content now available" after "no details yet" detection
This function previously requested each security content before
determining it is not the right one. It did this for many releases on
each run if there were any releases still with "no details yet".

Now it only reads data from the main page table as all needed data is
there.