ref+Twitter: make request manually, remove tweepy package
Tweepy does not seems to be actively developed. Last meaningful change
seems to be August 2023.

So for this small usage of login and post, it is also very simple to
make manual requests.

But why change as tweepy was working just fine? Well the package can
no longer run with Python 3.13 because of deprecation and removal of
some builtin packages (imghdr), so this prompted to look at the project
and see old issues about this with no response and no recent commits.