fix+BlueSky: remove new line character from the end of the posts
Some social media like Twitter and Mastodon seem to cut new line
characters from the end of the post and if the middle contains multiple
new lines, it will be trimmed to just one empty line.

BlueSky (at least via API) does not behave like this, posts are as sent.

So this commit should bring BlueSky in line with others while not
changing others.