fix: save more releases then the number of checked releases
Yesterday a weird ordering issue happened. After the GarageBand update,
the bot checked last 20 releases against the saved/posted 20 releases.
It figured the 20th one was not posted. While posting the 20th one and
removing the last one, at the next check, the last one was again "new".
This repeated a few times until a certain release.

I first thought that Apple must have reordered the table and caused this
as by checking the saved releases list, visionOS 2.2 was not at
the right location and that is the release whose removal by GarageBand
started the issue. It stopped at the element before where visionOS 2.2
should be.

It was not in fact Apple's fault as Internet Archive showed no
difference since December! I then went back to December in what order
visionOS 2.2 was posted and saved... I found that it was not posted and
while fixing it, manually saving it, I saved it at the end instead
of where it should be! :')

Enough of rambling. I have increased the number of saved items to 30
while leaving the number of checked to 20. This way in case of a few mine
or Apples reorderings, it should still be okay!